Functions to extract prediction intervals and observed data

Functions to extract prediction intervals and observed data

Usage

extract_hr_pi(hr.pi, outtype = c("long", "wide"))

extract_km_pi(km.pi, trunc.sim.censor = TRUE)

extract_medsurv_pi(km.pi, outtype = c("long", "wide"))

extract_medsurv_delta_pi(km.pi, outtype = c("long", "wide"))

Arguments

hr.pi

a return object from calc_hr_pi() function.

outtype

Specifies whether output will be in long or wide format.

km.pi

A return object from calc_km_pi() function.

trunc.sim.censor

A logical specifying whether to truncate the simulated curve at the last time of censor.dur specified in surv_param_sim().

Details

extract_hr_pi() extracts prediction intervals of simulated hazard ratios and the corresponding observed values.

extract_km_pi() extracts prediction intervals of simulated Kaplan-Meier curves.

extract_medsurv_pi() extracts prediction intervals of median survival times and and the corresponding observed values.

extract_medsurv_delta_pi() extracts prediction intervals of delta of median survival times between treatment groups